flask-backend/flask-server/requirements.txt
ultralytics==8.2.57
flask==3.0.3
Flask-Cors==4.0.1
transformers==4.42.4
pillow==10.4.0
pillow-heif==0.18.0
torch==2.3.1+cu121 --index-url https://download.pytorch.org/whl/cu121
torchvision==0.18.1 --index-url https://download.pytorch.org/whl/cu121

